AAC-first redesign: unified Composer + compact phrase grid
Replace TypingDock (706 lines) and TypingArea (622 lines) with a
single Composer component (614 lines) that works on both mobile and
desktop. Net reduction of 1212 lines.

Composer uses a compact single-line input strip with auto-grow (max
2 rows) and an expand button for longer messages. Action buttons
(Clear, Fix Text, Live Typing) appear in a secondary row only when
text exists.

Phrase grid optimized for AAC:
- Remove aspect-square from tiles, use min-h-[52px] compact rectangles
- Mobile: 2 → 3 columns, Desktop: up to 6 columns
- Replace badge indicators with left-border indicators
- Mobile visible phrases: 4-6 → ~12-15 without scrolling

SpeakButton unified: remove variant prop, single compact style.

Adjust --bottom-stack-height CSS variable for thinner Composer strip.

Closes #355
